requirements.txt
netCDF4
psycopg2-binary
python-dotenv